refactor: extract ImgMath
